Applicant Refactoring; RequestBody로 요청 변경 및 패키지 분리
