+++ gomen/src/main/java/com/tuna/gomen/applicant/command/entity/Applicant.java
@@ -1,4 +1,4 @@
-package com.tuna.gomen.applicant.entity;
+package com.tuna.gomen.applicant.command.entity;
 
 import com.tuna.gomen.teamRecruitBoard.entity.TeamRecruitBoard;
 import com.tuna.gomen.user.command.entity.User;
