+++ gomen/src/main/java/com/tuna/gomen/applicant/command/repository/ApplicantRepository.java
@@ -1,7 +1,7 @@
-package com.tuna.gomen.applicant.repository;
+package com.tuna.gomen.applicant.command.repository;
 
-import com.tuna.gomen.applicant.entity.Applicant;
-import com.tuna.gomen.applicant.entity.ApplicantId;
+import com.tuna.gomen.applicant.command.entity.Applicant;
+import com.tuna.gomen.applicant.command.entity.ApplicantId;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
