+++ gomen/src/main/java/com/tuna/gomen/applicant/command/service/ApplicantCommandService.java
@@ -1,0 +1,7 @@
+package com.tuna.gomen.applicant.command.service;
+
+public interface ApplicantCommandService {
+    void applyToTeam(Integer teamPostId, Integer userId, String introduce);
+    void cancelApplication(Integer teamPostId, Integer userId);
+    void processApplication(Integer teamPostId, Integer userId, String isProcessed);
+}
