+++ gomen/src/main/java/com/tuna/gomen/teamRecruitBoard/dto/TeamRecruitWithApplicantsDTO.java
@@ -1,14 +1,10 @@
 package com.tuna.gomen.teamRecruitBoard.dto;
 
-import com.tuna.gomen.applicant.dto.ApplicantDTO;
-import com.tuna.gomen.user.command.entity.User;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-
-import java.util.List;
 
 @Getter
 @Setter
